requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
black==25.1.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Crawl4AI==0.6.3
cryptography==45.0.5
cssselect==1.3.0
distro==1.9.0
fake-http-header==0.3.5
fake-useragent==2.2.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0

greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.5
hiredis==3.2.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.2
humanize==4.12.3
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1

lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
multidict==6.6.3
mypy_extensions==1.1.0
nltk==3.9.1
numpy==2.3.1

packaging==25.0
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
playwright==1.53.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
python-dotenv==1.1.1
python-ulid==1.1.0
PyYAML==6.0.2
rank-bm25==0.2.2
redis==5.3.0
redis-om==0.3.5
gptcache==0.1.44
fastapi==0.115.0
uvicorn==0.30.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
tf-playwright-stealth==1.2.0
tiktoken==0.9.0
tokenizers==0.21.2
tqdm==4.67.1
types-cffi==1.17.0.20250523
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==80.9.0.20250529
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
